refactor(gilgen): move GILGenModule to its own file
